Use C_ERROR instead of -1 in DetourPatcher.cpp, add new length-disassembler

+++ src/JKA_YBEProxy/DetourPatcher/DetourPatcher.hpp
@@ -11,10 +11,10 @@
 unsigned char	*Attach( unsigned char *pAddress, unsigned char *pNewAddress );
 void			 DisAssemble( unsigned char *iptr0, unsigned int *osizeptr );
 unsigned char	*Detach( unsigned char *pAddress, unsigned char *pTramp );
-unsigned int	 GetLen( unsigned char *pAddress );
-unsigned char	*GetTramp( unsigned char *pAddress, unsigned int iLen );
+size_t			 GetLen( unsigned char *pAddress );
+unsigned char	*GetTramp( unsigned char *pAddress, size_t iLen );
 unsigned int	 InlineFetch( unsigned char *pAddress );
 unsigned int	 InlinePatch( unsigned char *pAddress, unsigned char *pNewAddress );
 void			 Patch( unsigned char *pAddress, unsigned char bByte );
-void			 ReProtect( void *pAddress, int iLen );
-void			 UnProtect( void *pAddress, int iLen );
+void			 ReProtect( void *pAddress, size_t iLen );
+void			 UnProtect( void *pAddress, size_t iLen );
